Move constants, utils, main to airbrakes package
Also add ability to run with uvx
