+++ scripts/run_imu.py
@@ -4,7 +4,7 @@ Make sure you are in the root directory of the project, not inside scripts, and 
 For the pi, you will have to use python3
 """
 
-from constants import IMU_PORT
+from airbrakes.constants import IMU_PORT
 from airbrakes.hardware.imu import IMU
