+++ scripts/run_logger.py
@@ -4,7 +4,7 @@ import time
 from collections import deque
 
 from airbrakes.data_handling.data_processor import IMUDataProcessor
-from constants import TEST_LOGS_PATH
+from airbrakes.constants import TEST_LOGS_PATH
 from airbrakes.data_handling.imu_data_packet import RawDataPacket
 from airbrakes.data_handling.logger import Logger
