+++ src/airbrakes/hardware/servo.py
@@ -6,7 +6,7 @@ import warnings
 
 import gpiozero
 
-from constants import SERVO_DELAY_SECONDS, ServoExtension
+from airbrakes.constants import SERVO_DELAY_SECONDS, ServoExtension
 
 
 class Servo:
