+++ src/airbrakes/main.py
@@ -7,6 +7,7 @@ import time
 from gpiozero.pins.mock import MockFactory, MockPWMPin
 
 from airbrakes.airbrakes import AirbrakesContext
+from airbrakes.constants import IMU_PORT, LOGS_PATH, SERVO_PIN
 from airbrakes.data_handling.apogee_predictor import ApogeePredictor
 from airbrakes.data_handling.data_processor import IMUDataProcessor
 from airbrakes.data_handling.logger import Logger
@@ -15,11 +16,37 @@ from airbrakes.hardware.servo import Servo
 from airbrakes.mock.display import FlightDisplay
 from airbrakes.mock.mock_imu import MockIMU
 from airbrakes.mock.mock_logger import MockLogger
-from constants import IMU_PORT, LOGS_PATH, SERVO_PIN
-from utils import arg_parser
+from airbrakes.utils import arg_parser
 
 
-def create_components(args: argparse.Namespace) -> tuple[Servo, IMU, Logger]:
+def run_real_flight() -> None:
+    """Entry point for the application to run the real flight."""
+    args = arg_parser()
+    run_flight(args)
+
+
+def run_mock_flight() -> None:
+    """Entry point for the application to run the mock flight."""
+    args = arg_parser()
+    args.mock = True
+    run_flight(args)
+
+
+def run_flight(args: argparse.Namespace) -> None:
+    sim_time_start = time.time()
+
+    servo, imu, logger, data_processor, apogee_predictor = create_components(args)
+    # Initialize the airbrakes context and display
+    airbrakes = AirbrakesContext(servo, imu, logger, data_processor, apogee_predictor)
+    flight_display = FlightDisplay(airbrakes, sim_time_start, args)
+
+    # Run the main flight loop
+    run_flight_loop(airbrakes, flight_display, args.mock)
+
+
+def create_components(
+    args: argparse.Namespace,
+) -> tuple[Servo, IMU, Logger, IMUDataProcessor, ApogeePredictor]:
     """
     Creates the system components needed for the airbrakes system. Depending on its arguments, it
     will return either mock or real components.
@@ -45,7 +72,10 @@ def create_components(args: argparse.Namespace) -> tuple[Servo, IMU, Logger]:
         imu = IMU(IMU_PORT)
         logger = Logger(LOGS_PATH)
 
-    return servo, imu, logger
+    # Initialize data processing and prediction
+    data_processor = IMUDataProcessor()
+    apogee_predictor = ApogeePredictor()
+    return servo, imu, logger, data_processor, apogee_predictor
 
 
 def run_flight_loop(
@@ -84,36 +114,14 @@ def run_flight_loop(
         airbrakes.stop()
 
 
-def main(args: argparse.Namespace) -> None:
-    """
-    Sets up and runs the airbrakes system.
-    :param args: Command line arguments for configuration.
-    """
-    sim_time_start = time.time()
-
-    # Creates the components for the airbrakes system
-    servo, imu, logger = create_components(args)
-
-    # Initialize data processing and prediction
-    data_processor = IMUDataProcessor()
-    apogee_predictor = ApogeePredictor()
-
-    # Initialize the airbrakes context and display
-    airbrakes = AirbrakesContext(servo, imu, logger, data_processor, apogee_predictor)
-    flight_display = FlightDisplay(airbrakes, sim_time_start, args)
-
-    # Run the main flight loop
-    run_flight_loop(airbrakes, flight_display, args.mock)
-
-
-if __name__ == "__main__":
-    # Command line args (after these are run, you can press Ctrl+C to exit the program):
-    # python main.py -v: Shows the display with much more data
-    # python main.py -m: Runs a simulation on your computer
-    # python main.py -m -r: Runs a simulation on your computer with the real servo
-    # python main.py -m -l: Runs a simulation on your computer and keeps the log file after the
-    # simulation stops
-    # python main.py -m -f: Runs a simulation on your computer at full speed
-    # python main.py -m -d: Runs a simulation on your computer in debug mode (doesn't show display)
-    args = arg_parser()
-    main(args)
+# if __name__ == "__main__":
+#     # Command line args (after these are run, you can press Ctrl+C to exit the program):
+#     # uvx real -v: Shows the display with much more data
+#     # uvx mock: Runs a simulation on your computer
+#     # uvx mock -r: Runs a simulation on your computer with the real servo
+#     # uvx mock -l: Runs a simulation on your computer and keeps the log file after the
+#     # simulation stops
+#     # uvx mock -f: Runs a simulation on your computer at full speed
+#     # uvx mock -d: Runs a simulation on your computer in debug mode (doesn't show display)
+#     args = arg_parser()
+#     main(args)
