+++ tests/test_imu.py
@@ -7,13 +7,13 @@ from collections import deque
 import faster_fifo
 import pytest
 
+from airbrakes.constants import IMU_PORT, STOP_SIGNAL
 from airbrakes.data_handling.imu_data_packet import (
     EstimatedDataPacket,
     IMUDataPacket,
     RawDataPacket,
 )
 from airbrakes.hardware.imu import IMU
-from constants import IMU_PORT, STOP_SIGNAL
 
 
 class TestIMU:
